requirements.txt
amqp==5.1.1
aniso8601==9.0.1
apturl==0.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
bcrypt==4.0.0
billiard==4.1.0
blinker==1.6.2
Brlapi==0.7.0
Brotli==1.0.9
cachelib==0.9.0
celery==5.3.4
certifi==2019.11.28
cffi==1.15.1
chardet==5.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
contourpy==1.0.5
cryptography==2.8
cssselect2==0.7.0
cycler==0.11.0
dbus-python==1.2.16
defer==1.0.6
Deprecated==1.2.13
distlib==0.3.7
distro==1.4.0
dnspython==2.2.1
email-validator==1.2.1
entrypoints==0.3
excel==1.0.0
fasteners==0.14.1
filelock==3.12.3
Flask==2.1.0
Flask-Cache==0.13.1
Flask-Caching==2.0.1
Flask-Cors==3.0.10
Flask-Excel==0.0.7
Flask-Ext==0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==5.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.37.3
future==0.18.2
greenlet==1.1.3
html5lib==1.1
httplib2==0.14.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
keyring==18.0.1
kiwisolver==1.4.4
kombu==5.3.2
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
MarkupSafe==2.1.1
matplotlib==3.6.0
monotonic==1.5
netifaces==0.10.4
numpy==1.23.3
oauthlib==3.1.0
olefile==0.46
packaging==21.3
pandas==2.0.3
paramiko==2.6.0
passlib==1.7.4
pexpect==4.6.0
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.31
protobuf==3.6.1
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydyf==0.3.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
PyGObject==3.36.0
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
pyphen==0.13.0
pyRFC3339==1.1
python-dateutil==2.8.2
pytz==2022.2.1
pyxdg==0.26
PyYAML==5.3.1
redis==4.3.4
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
SQLAlchemy==1.4.40
texttable==1.6.4
tinycss2==1.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.25.8
vine==5.0.0
virtualenv==20.24.3
wadllib==1.3.3
wcwidth==0.2.5
weasyprint==56.1
webencodings==0.5.1
Werkzeug==2.3.7
wrapt==1.14.1
WTForms==3.0.1
xlrd==2.0.1
zipp==3.8.1
zopfli==0.2.1
